Clean Code + Fetch au démarage
