+++ InstaGrans/app/src/main/java/edu/mj102660/instagrans/MainActivity.java
@@ -3,7 +3,7 @@ package edu.mj102660.instagrans;
 import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.Intent;
-import android.os.AsyncTask;
+
 import android.os.Bundle;
 import android.widget.EditText;
 import android.widget.ImageButton;
@@ -34,18 +34,8 @@ import edu.mj102660.instagrans.profile.ProfileActivity;
 public class MainActivity extends AppCompatActivity implements ClickableActivity {
 
     private ActivityMainBinding binding;
-
     BottomNavigationView navView;
-    ImageButton searchButton;
-    EditText searchText;
-    ProgressDialog pDialog ;
-    private ListView lv;
 
-    //todo Il faut impérativement renseigner l'ip du webservice ici
-    private final String hostIp = "192.168.0.35";
-
-    Intent intent;
-    ArrayList<Granny> grannies = new ArrayList<>();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -61,13 +51,7 @@ public class MainActivity extends AppCompatActivity implements ClickableActivity
         NavigationUI.setupWithNavController(binding.navView, navController);
 
 
-        //Fetch les données du JSon
-        new GetGrannies().execute();
-        try {
-            Thread.sleep(1000);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
+
 
     }
 
@@ -89,116 +73,6 @@ public class MainActivity extends AppCompatActivity implements ClickableActivity
 
 
 
-    /**
-     * Tache asynchrone pour fetch
-     */
-    public class GetGrannies extends AsyncTask<Void,Void,Void> {
 
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            pDialog = new ProgressDialog(MainActivity.this);
-            pDialog.setMessage("Connexion en cours...");
-            pDialog.setCancelable(false);
-            pDialog.show();
-
-        }
-
-
-
-        @Override
-        // appeler automatiquement après onPreExecute
-        protected Void doInBackground(Void... params) {
-
-            for (int progress = 0; progress  < 1000000000; progress++){
-                // Ne fait rien mais fait juste passer du temps
-            }
-            HttpHandler handler = new HttpHandler();
-
-            try {
-                System.out.println(InetAddress.getLocalHost());
-            } catch (UnknownHostException e) {
-                throw new RuntimeException(e);
-            }
-            //192.168.0.35
-            String jsonStr = handler.makeServiceCall("http://"+hostIp+":8080/getGrannies");
-
-            if (jsonStr != null) {
-                try {
-                    JSONObject jsonObj = new JSONObject(jsonStr);
-
-                    //Récupérer l'array
-                    JSONArray grannyList = jsonObj.getJSONArray("grannies");
-
-                    // pour chaque element de l'array
-                    for (int i = 0; i < grannyList.length(); i++) {
-                        JSONObject c = grannyList.getJSONObject(i);
-
-
-                        //Récupération des attributs
-                        String name= c.getString("name");
-                        int age = c.getInt("age");
-                        String location = c.getString("location");
-                        String desc = c.getString("desc");
-                        String urlPicture = c.getString("urlPicture");
-                        double score = c.getDouble("score");
-                        double price = c.getDouble("price");
-
-                        //On créer l'objet, on applique tout les attributs
-                        Granny granny = new Granny();
-                        granny.setName(name);
-                        granny.setAge(age);
-                        granny.setLocation(location);
-                        granny.setDesc(desc);
-                        granny.setScore(score);
-                        granny.setPrice(price);
-                        granny.setUrlPicture(urlPicture);
-
-
-                        //Récupération de l'array
-                        JSONArray dishesArray = c.getJSONArray("dishes");
-                        for (int j =0; j<dishesArray.length();j++){
-
-                            //Récupération de l'objet Dish
-                            JSONObject d = dishesArray.getJSONObject(j);
-                            Dish dish = new Dish();
-                            dish.setName(d.getString("name"));
-                            //System.out.println("name  "+d.getString("name"));
-                            dish.setNote(d.getString("note"));
-                            dish.setUrlImage(d.getString("urlImage"));
-                            dish.setPrepTime(d.getString("prepTime"));
-                            dish.setPrepMinute(d.getInt("prepMinute"));
-                            granny.addDish(dish);
-                        }
-
-                        //Ajout de la granny au SINGLETON
-                        Grans.getInstance().add(granny);
-                    }
-                    System.out.println("Grannies Fetched.");
-
-                } catch (final JSONException e) {
-                    System.out.println("Erreur JSON " + e.getMessage());
-
-                }
-            } else {
-                System.out.println( "Probleme connexion ");
-            }
-
-            return null;
-        }
-
-
-        @Override
-        protected void onPostExecute(Void result) {
-            super.onPostExecute(result);
-
-            if (pDialog.isShowing())
-                pDialog.dismiss();
-            ArrayList<String> list = new ArrayList<String>();
-            for(int i=0;i<grannies.size();i++)
-                list.add(grannies.get(i).toString());
-        }
-
-    }
 
 }
